% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{atlasIntroServer}
\alias{atlasIntroServer}
\title{Server logic for atlas instructions}
\usage{
atlasIntroServer(id, config_file = "config.yaml", config_id)
}
\description{
Server logic for atlas instructions
}
