% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{comparaIntroServer}
\alias{comparaIntroServer}
\title{Server logic for comparative instructions}
\usage{
comparaIntroServer(id, config_file = "config.yaml", config_id)
}
\description{
Server logic for comparative instructions
}
