% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{comparaServer}
\alias{comparaServer}
\title{Server logic for cross-species comparison}
\usage{
comparaServer(id, config_file = "config.yaml", config_id1, config_id2)
}
\description{
Server logic for cross-species comparison
}
