% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dictionary_t2g}
\alias{dictionary_t2g}
\title{Create a transcript to gene dictionary from a GTF annotation file}
\usage{
dictionary_t2g(
  gtf_fn,
  vector_to_fix,
  t2g = TRUE,
  transcript_field = "transcript",
  transcript_id = "transcript_id",
  gene_id = "gene_id",
  return_elements_not_in_gtf = TRUE
)
}
\description{
Create a transcript to gene dictionary from a GTF annotation file
}
