% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fread_cell_annotation}
\alias{fread_cell_annotation}
\title{Import metacell annotations}
\usage{
fread_cell_annotation(file, select = NULL)
}
\arguments{
\item{file}{path to file to load}
}
\value{
daata.table with the following columns: mc, cell_type, color
}
\description{
Import metacell annotations
}
