% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mc_2d_plot}
\alias{mc_2d_plot}
\title{Plot 2d proj colored by mc}
\usage{
mc_2d_plot(
  mc2d,
  mcsc,
  cttable,
  plot_edges = TRUE,
  plot_mcs = TRUE,
  plot_mc_name = TRUE,
  width = 12,
  height = 12,
  res = NA,
  cex_mc = 3,
  cex_sc = 0.75
)
}
\description{
Plot 2d proj colored by mc
}
