% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mc_gene_summary}
\alias{mc_gene_summary}
\title{Generate table for highly expressed genes in a group of metacells}
\usage{
mc_gene_summary(
  mc_ids,
  fc = 2,
  method = c("absolute", "median"),
  methodbg = method,
  lky = 0,
  usefcbg = FALSE,
  fcbg = fc,
  lkybg = lky,
  mc_fp,
  mc_counts,
  annt,
  tfannt
)
}
\arguments{
\item{method, }{method to use for retrieving gnees expressed in metacells,
either "absolute" or "median"; if "absolute", all metacells (- lky percentage)
must have gene fc above threshold; if "median", the median in all metacells
must be above threshold}
}
\description{
Generate table for highly expressed genes in a group of metacells
}
