% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mgenes_hmap_height}
\alias{mgenes_hmap_height}
\title{Plot multigene heatmap to show expression of selected genes}
\usage{
mgenes_hmap_height(
  nmat,
  gids,
  annt,
  min_expression_fc = NULL,
  max_expression_fc = NULL
)
}
\description{
Plot multigene heatmap to show expression of selected genes
}
