% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{multiGeneServer}
\alias{multiGeneServer}
\title{Server logic for visualizing multiple gene expression}
\usage{
multiGeneServer(id, config_file = "config.yaml", config_id)
}
\description{
Server logic for visualizing multiple gene expression
}
