% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{multiGeneUI}
\alias{multiGeneUI}
\title{User interface for visualizing multiple gene expression}
\usage{
multiGeneUI(id, label = "Multi gene expression")
}
\description{
User interface for visualizing multiple gene expression
}
