% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{orthgServer}
\alias{orthgServer}
\title{Server logic for homologos across species}
\usage{
orthgServer(id, config_file = "config.yaml", sps = NULL)
}
\description{
Server logic for homologos across species
}
