% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scp_plot_cmod_markers_mc}
\alias{scp_plot_cmod_markers_mc}
\title{Plot heatmap of gene expression for metacells}
\usage{
scp_plot_cmod_markers_mc(
  marker_data_list,
  output_file = NULL,
  height = 10,
  width = 5,
  res = NA,
  show_heatmap_legend = TRUE,
  highlight_genes = NULL,
  show_gene_names = FALSE,
  gene_font_size = 5,
  clust_col = NULL,
  clust_bars = NULL,
  clust_col_others = NULL,
  col_others = NULL,
  clust_anno_size = unit(1, "mm"),
  show_mc_names = TRUE,
  mc_font_size = 5,
  heatmap_colors = c("white", "gray99", "orange", "orangered2", "#520c52"),
  max_expression_fc = 5,
  gene_chr_limit = 70,
  verbose = FALSE,
  save_rds = TRUE,
  print_border = TRUE,
  show_clust_borders = TRUE
)
}
\description{
Plot heatmap of gene expression for metacells
}
