% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scp_plot_cmod_markers_sc}
\alias{scp_plot_cmod_markers_sc}
\title{Plot heatmap of gene expression for single cells}
\usage{
scp_plot_cmod_markers_sc(
  marker_data_list,
  mcsc,
  umat,
  output_file = NULL,
  height = 10,
  width = 5,
  res = NA,
  show_heatmap_legend = TRUE,
  highlight_genes = NULL,
  show_gene_names = FALSE,
  gene_font_size = 5,
  clust_col = NULL,
  clust_bars = NULL,
  clust_col_others = NULL,
  col_others = NULL,
  clust_anno_size = unit(1, "mm"),
  show_mc_names = TRUE,
  mc_font_size = 5,
  heatmap_colors = c("white", "white", "orange", "red", "purple", "black"),
  gene_chr_limit = 70,
  verbose = FALSE,
  save_rds = FALSE,
  smoothen = 5,
  max_expression_fc = 5,
  max_expression_fc_sc = 5,
  print_border = TRUE,
  show_clust_borders = TRUE
)
}
\description{
Plot heatmap of gene expression for single cells
}
