% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scp_plot_sc_2d_gene_exp}
\alias{scp_plot_sc_2d_gene_exp}
\title{Plot 2d expression of a gene}
\usage{
scp_plot_sc_2d_gene_exp(
  mc2d,
  nmat,
  umat,
  sid,
  annt,
  gene_id = NULL,
  sc_vector = NULL,
  sc_scale = "unidir",
  sc_transform = NULL,
  sc_min = 0,
  sc_max = 8,
  sc_label = "cell UMI",
  sc_zero_color = "gray95",
  do_umifrac_sc = FALSE,
  mc_vector = NULL,
  mc_scale = "unidir",
  mc_transform = log2,
  mc_min = 0,
  mc_max = 2,
  mc_label = "log2(FC)",
  mc_zero_color = NULL,
  unidir_color_scale = c("gray90", "orange", "orangered2", "#520c52"),
  bidir_color_scale = c("midnightblue", "dodgerblue3", "deepskyblue", "#b8e0ed",
    "gray90", "#eccac0", "#ff8d36", "#f34312", "#8e0631"),
  plot_edges = TRUE,
  plot_mcs = TRUE,
  plot_mc_name = TRUE,
  width = 12,
  height = 12,
  res = NA,
  cex_mc = 3,
  cex_sc = 0.75,
  do_axes = FALSE
)
}
\arguments{
\item{mc2d}{mc2d object}

\item{nmat}{mc_fp matrix}

\item{umat}{sc UMI matrix}
}
\description{
Plot 2d expression of a gene
}
