% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sg_plot}
\alias{sg_plot}
\title{Plot an expression barplots of a gene}
\usage{
sg_plot(
  nmat,
  umat,
  cttable,
  gid = NULL,
  sid = NULL,
  mdnorm,
  annt,
  ctpalette = NULL
)
}
\arguments{
\item{nmat}{mc_fp matrix}

\item{umat}{sc UMI matrix}
}
\description{
Plot an expression barplots of a gene
}
