% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{singleGeneServer}
\alias{singleGeneServer}
\title{Server logic for visualizing single gene expression}
\usage{
singleGeneServer(id, config_file = "config.yaml", config_id)
}
\description{
Server logic for visualizing single gene expression
}
