% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{singleGeneUI}
\alias{singleGeneUI}
\title{User interface for visualizing single gene expression}
\usage{
singleGeneUI(id, label = "Single gene expression")
}
\description{
User interface for visualizing single gene expression
}
