% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{summaryServer}
\alias{summaryServer}
\title{Server logic for summarizing expression in group of metacells}
\usage{
summaryServer(id, config_file = "config.yaml", config_id)
}
\description{
Server logic for summarizing expression in group of metacells
}
