% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{summaryUI}
\alias{summaryUI}
\title{User interface for summarizing expression in group of metacells}
\usage{
summaryUI(id, config_file = "config.yaml", label = "Metacell summary")
}
\description{
User interface for summarizing expression in group of metacells
}
