% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sg_plot}
\alias{sg_plot}
\title{Plot an expression barplots of a gene}
\usage{
sg_plot(
  nmat,
  umat = NULL,
  cttable,
  gene_id,
  gid = NULL,
  sid = NULL,
  mdnorm = FALSE,
  annt,
  order_by = "metacell",
  ctpalette = NULL,
  legend.position = "none",
  mc_label_size = 9,
  title = FALSE,
  caption = "Dashed line indicates the maximum observed value."
)
}
\arguments{
\item{nmat}{mc_fp matrix}

\item{umat}{metacell UMI matrix}

\item{cttable}{metacell annotation table}

\item{mdnorm}{normalize?}

\item{annt}{gene annotations table}

\item{order_by}{\code{"metacell"} (numerically) or \code{"cell_type"} (as ordeerd in the annotation file)}

\item{ctpalette}{custom colours; if \code{NULL}, colors taken from metacell anotation file}
}
\description{
Plot an expression barplots of a gene
}
